requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-environ==0.11.2
djangorestframework==3.15.1
Faker==24.14.0
pillow==10.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-environ==0.4.54
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1